pyproject.toml
[project]
name = "datia-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "debugpy>=1.8.12",
    "fastapi>=0.115.8",
    "langchain-openai>=0.3.6",
    "langgraph>=0.2.74",
    "langgraph-cli[inmem]>=0.1.72",
    "notebook>=7.3.2",
    "pandas>=2.2.3",
    "pydantic-ai>=0.0.24",
    "pydantic-settings>=2.7.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
